Improve handling of adding items to empty playlist in MediaController
This is a follow-up to https://github.com/androidx/media/commit/415ad5a33c4d67f0ced7411686f3949f22c077b1 where we made the same change in
ExoPlayerImpl and SimpleBasePlayer, but for consistency it makes
sense to also update the masking code in MediaControllerImplBase to
assume the same logic.

Note: MediaControllerImplLegacy already handles this case via
setMediaItems and doesn't need to be updated further.

#minor-release

PiperOrigin-RevId: 534038759