Update `WebvttParser` to implement new partial-output method
This requires adapting the 'to `CuesWithTiming` list' logic to work with
the new partial-output API, and that needs a private method so it's no
longer a good fit for a default method on `Subtitle` - hence moving it
to a new utility class.

Also update the implementation to never return `UNSET` duration (this is
an equivalent change to the `SsaParser` change in https://github.com/androidx/media/commit/ebd5348d38541f0c6b8e040f1c7cdcd559ea4a7a).

PiperOrigin-RevId: 566598094